:sparkles: Add mount option in features
